+++ Drivers/Super/driver_super.h
@@ -26,7 +26,7 @@
 #include <src_user/hal/if_list.h>
 #include <stdint.h>
 #include "../../library/endian.h"        // パスが不定な自動生成コード類で使えるように
-#include "../../System/TimeManager/time_manager.h"
+#include "../../system/time_manager/time_manager.h"
 
 #define DS_STREAM_MAX                 (3)         /*!< DS_StreamConfig の最大数
                                                        uint8_t を想定          */
