+++ TlmCmd/block_command_loader.h
@@ -10,7 +10,7 @@
 
 #include "./block_command_table.h"
 #include "../applications/timeline_command_dispatcher_id_define.h"
-#include "../System/TimeManager/obc_time.h"
+#include "../system/time_manager/obc_time.h"
 #include <src_user/applications/app_registry.h>
