+++ TlmCmd/common_cmd_packet.h
@@ -6,7 +6,7 @@
 #ifndef COMMON_CMD_PACKET_H_
 #define COMMON_CMD_PACKET_H_
 
-#include "../System/TimeManager/obc_time.h"
+#include "../system/time_manager/obc_time.h"
 #include <src_user/TlmCmd/command_definitions.h>
 
 // ここで CCP_DEST_TYPE を定義する
