+++ TlmCmd/common_tlm_packet.h
@@ -6,7 +6,7 @@
 #ifndef COMMON_TLM_PACKET_H_
 #define COMMON_TLM_PACKET_H_
 
-#include "../System/TimeManager/obc_time.h"
+#include "../system/time_manager/obc_time.h"
 #include <src_user/TlmCmd/telemetry_definitions.h>
 
 // ここで， CTP_APID_FROM_ME, CTP_MAX_LEN, CommonTlmPacket として使うパケット型を指定する
