+++ TlmCmd/packet_list.c
@@ -5,7 +5,7 @@
  */
 #include "packet_list.h"
 
-#include "../System/TimeManager/time_manager.h"
+#include "../system/time_manager/time_manager.h"
 #include "common_tlm_cmd_packet.h"
 #include "block_command_executor.h"
 #include "block_command_table.h"
