+++ applications/divided_cmd_utility.h
@@ -13,9 +13,9 @@
 #ifndef DIVIDED_CMD_UTILITY_H_
 #define DIVIDED_CMD_UTILITY_H_
 
-#include "../System/ApplicationManager/app_info.h"
+#include "../system/application_manager/app_info.h"
 #include "../TlmCmd/common_cmd_packet.h"
-#include "../System/TimeManager/obc_time.h"
+#include "../system/time_manager/obc_time.h"
 #include <src_user/TlmCmd/command_definitions.h>
 
 #define DCU_LOG_MAX    (16)    //!< 保存するログの最大数
