+++ applications/event_utility.h
@@ -6,7 +6,7 @@
 #ifndef EVENT_UTILITY_H_
 #define EVENT_UTILITY_H_
 
-#include "../System/ApplicationManager/app_info.h"
+#include "../system/application_manager/app_info.h"
 #include "../TlmCmd/common_cmd_packet.h"
 
 AppInfo EVENT_UTIL_create_app(void);
