+++ applications/timeline_command_dispatcher.c
@@ -1,8 +1,8 @@
 #pragma section REPRO
 #include "timeline_command_dispatcher.h"
 #include "../TlmCmd/packet_handler.h"
-#include "../System/TimeManager/time_manager.h"
-#include "../System/EventManager/event_logger.h"
+#include "../system/time_manager/time_manager.h"
+#include "../system/event_manager/event_logger.h"
 #include "../TlmCmd/common_cmd_packet_util.h"
 #include "../library/result.h"
