+++ applications/utility_command.h
@@ -5,7 +5,7 @@
 
 #include "../TlmCmd/common_cmd_packet.h"
 #include "../hal/uart.h"
-#include "../System/ApplicationManager/app_info.h"
+#include "../system/application_manager/app_info.h"
 
 #define UTIL_CMD_SIZE_MAX (256) // 256Bytes以上コマンドとして送ることはないはず．．．ICOUPS_PARAM_BLより長いし
