+++ c2a_core_main.c
@@ -3,12 +3,12 @@
 
 #include "./library/git_revision.h"
 #include "./library/print.h"
-#include "./System/TaskManager/task_dispatcher.h"
-#include "./System/ApplicationManager/app_manager.h"
-#include "./System/EventManager/event_manager.h"
-#include "./System/TimeManager/time_manager.h"
-#include "./System/ModeManager/mode_manager.h"
-#include "./System/WatchdogTimer/watchdog_timer.h"
+#include "./system/task_manager/task_dispatcher.h"
+#include "./system/application_manager/app_manager.h"
+#include "./system/event_manager/event_manager.h"
+#include "./system/time_manager/time_manager.h"
+#include "./system/mode_manager/mode_manager.h"
+#include "./system/watchdog_timer/watchdog_timer.h"
 #include "./TlmCmd/packet_handler.h"
 #include "./TlmCmd/block_command_table.h"
 #include "./TlmCmd/command_analyze.h"
